requirements.txt
Appium-Python-Client==3.1.1
attrs==23.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==0.13.1
py==1.11.0
PySocks==1.7.1
pytest==7.4.3
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
